requirements.txt
dialogflow>=1.0.0
python-telegram-bot>=12.8
SQLAlchemy==1.3.18
